// prisma/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UnitPreference = {
  kg: 'kg',
  lbs: 'lbs'
} as const

export type UnitPreference = (typeof UnitPreference)[keyof typeof UnitPreference]


export const Gender = {
  male: 'male',
  female: 'female',
  other: 'other'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const SessionLabel = {
  Chest: 'Chest',
  Back: 'Back',
  Shoulders: 'Shoulders',
  Arms: 'Arms',
  Core: 'Core',
  Legs: 'Legs',
  Glutes: 'Glutes',
  FullBody: 'FullBody',
  Cardio: 'Cardio',
  Mobility: 'Mobility',
  Stretching: 'Stretching'
} as const

export type SessionLabel = (typeof SessionLabel)[keyof typeof SessionLabel]